Allow DVWindow to access the AVPlayer interface from C++ directly.
Instead of having to send signals to QML for video control.

+++ depthview2/include/dvinputinterface.hpp
@@ -46,5 +46,5 @@ public:
     virtual void playPauseVideo() = 0;
     virtual void seekBack() = 0;
     virtual void seekForward() = 0;
-    virtual void seekAmount(int msec) = 0;
+    virtual void seekAmount(qint64 msec) = 0;
 };
